+++ front/src/applications/editor/components/EntitySumUp.tsx
@@ -98,6 +98,7 @@ function getSumUpContent(
   let text = '';
   const subtexts: (string | JSX.Element)[] = [];
   const classes = { ...DEFAULT_CLASSES, ...(classesOverride || {}) };
+
   switch (entity.objType) {
     case 'TrackSection': {
       const trackSection = entity as TrackSectionEntity;
