+++ front/src/applications/editor/tools/rangeEdition/components.tsx
