+++ front/src/applications/editor/tools/rangeEdition/components/LpvPanelCard.tsx
