+++ front/src/applications/editor/tools/switchProps.ts
@@ -2,7 +2,7 @@ import { BufferStopEntity, DetectorEntity, SignalEntity } from 'types';
 import { RouteEditionState } from './routeEdition/types';
 import { SelectionState } from './selection/types';
 import { TrackEditionState } from './trackEdition/types';
-import { SpeedSectionEditionState } from './speedSectionEdition/types';
+import { SpeedSectionEditionState } from './rangeEdition/types';
 import { PointEditionState } from './pointEdition/types';
 import TOOL_TYPES from './toolTypes';
 import { SwitchEditionState } from './switchEdition/types';
