+++ front/src/applications/editor/tools/tools.ts
@@ -4,7 +4,7 @@ import {
   DetectorEditionTool,
   SignalEditionTool,
 } from './pointEdition/tools';
-import SpeedSectionEditionTool from './speedSectionEdition/tool';
+import { SpeedEditionTool } from './rangeEdition/tools';
 import TOOL_TYPES from './toolTypes';
 import SelectionTool from './selection/tool';
 import SwitchEditionTool from './switchEdition/tool';
@@ -13,7 +13,7 @@ import RouteEditionTool from './routeEdition/tool';
 const TOOLS = Object.freeze({
   [TOOL_TYPES.SELECTION]: SelectionTool,
   [TOOL_TYPES.TRACK_EDITION]: TrackEditionTool,
-  [TOOL_TYPES.SPEED_SECTION_EDITION]: SpeedSectionEditionTool,
+  [TOOL_TYPES.SPEED_SECTION_EDITION]: SpeedEditionTool,
   [TOOL_TYPES.SWITCH_EDITION]: SwitchEditionTool,
   [TOOL_TYPES.SIGNAL_EDITION]: SignalEditionTool,
   [TOOL_TYPES.DETECTOR_EDITION]: DetectorEditionTool,
